VCC/requirements.txt
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
prophet==1.1.4
plotly==5.18.0